add startServer and stopServer helpers to test suite

+++ test/test.js
@@ -7,14 +7,29 @@ global.log = function log() {
 
 before(function(done) {
   var port = 8001
-  log('starting target websocket server on port ' + port)
-  this.wsServer = fork(path.join(__dirname, 'ws-server.js'), [port])
+
+  this.startServer = function() {
+    log('starting test websocket server on port ' + port)
+    if (!this.wsServer) {
+      this.wsServer = fork(path.join(__dirname, 'ws-server.js'), [port])
+    }
+  }
+
+  this.stopServer = function() {
+    log('killing test websocket server')
+    if (this.wsServer) {
+      this.wsServer.kill()
+      this.wsServer = null
+    }
+  }
+
+  this.startServer()
+
   setTimeout(done, 1500)
 })
 
 after(function() {
-  log('killing websocket server')
-  this.wsServer.kill()
+  this.stopServer()
 })
 
 require('./single-socket')
